requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
databases
python-dotenv